+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ##### Fixed
 
+- Fixed delivery cost for order list
 - Improved order list rendering for administrator
 
 #### [2.0.1] - 2025-05-23